+++ src/CoreEx/Http/TypedHttpClientBase.cs
@@ -47,7 +47,11 @@ namespace CoreEx.Http
         /// <param name="args">Zero or more <see cref="IHttpArg"/> objects for <paramref name="requestUri"/> templating, query string additions, and content body specification.</param>
         /// <param name="cancellationToken">The <see cref="CancellationToken"/>.</param>
         /// <returns>The <see cref="HttpRequestMessage"/>.</returns>
+#if NET7_0_OR_GREATER
         protected Task<HttpRequestMessage> CreateRequestAsync(HttpMethod method, [StringSyntax(StringSyntaxAttribute.Uri)] string requestUri, HttpRequestOptions? requestOptions = null, IEnumerable<IHttpArg>? args = null, CancellationToken cancellationToken = default)
+#else
+        protected Task<HttpRequestMessage> CreateRequestAsync(HttpMethod method, string requestUri, HttpRequestOptions? requestOptions = null, IEnumerable<IHttpArg>? args = null, CancellationToken cancellationToken = default)
+#endif
             => CreateRequestInternalAsync(method, requestUri, null, requestOptions, args, cancellationToken);
 
         /// <summary>
@@ -60,7 +64,11 @@ namespace CoreEx.Http
         /// <param name="args">Zero or more <see cref="IHttpArg"/> objects for <paramref name="requestUri"/> templating, query string additions, and content body specification.</param>
         /// <param name="cancellationToken">The <see cref="CancellationToken"/>.</param>
         /// <returns>The <see cref="HttpRequestMessage"/>.</returns>
+#if NET7_0_OR_GREATER
         protected Task<HttpRequestMessage> CreateContentRequestAsync(HttpMethod method, [StringSyntax(StringSyntaxAttribute.Uri)] string requestUri, HttpContent content, HttpRequestOptions? requestOptions = null, IEnumerable<IHttpArg>? args = null, CancellationToken cancellationToken = default)
+#else
+        protected Task<HttpRequestMessage> CreateContentRequestAsync(HttpMethod method, string requestUri, HttpContent content, HttpRequestOptions? requestOptions = null, IEnumerable<IHttpArg>? args = null, CancellationToken cancellationToken = default)
+#endif
             => CreateRequestInternalAsync(method, requestUri, content, requestOptions, args, cancellationToken);
 
         /// <summary>
@@ -74,7 +82,11 @@ namespace CoreEx.Http
         /// <param name="args">Zero or more <see cref="IHttpArg"/> objects for <paramref name="requestUri"/> templating, query string additions, and content body specification.</param>
         /// <param name="cancellationToken">The <see cref="CancellationToken"/>.</param>
         /// <returns>The <see cref="HttpRequestMessage"/>.</returns>
+#if NET7_0_OR_GREATER
         protected Task<HttpRequestMessage> CreateJsonRequestAsync<TReq>(HttpMethod method, [StringSyntax(StringSyntaxAttribute.Uri)] string requestUri, TReq value, HttpRequestOptions? requestOptions = null, IEnumerable<IHttpArg>? args = null, CancellationToken cancellationToken = default)
+#else
+        protected Task<HttpRequestMessage> CreateJsonRequestAsync<TReq>(HttpMethod method, string requestUri, TReq value, HttpRequestOptions? requestOptions = null, IEnumerable<IHttpArg>? args = null, CancellationToken cancellationToken = default)
+#endif
             => CreateContentRequestAsync(method, requestUri, new StringContent(JsonSerializer.Serialize(value), Encoding.UTF8, MediaTypeNames.Application.Json), ApplyValueETagToRequestOptions(value, requestOptions), args, cancellationToken);
 
         /// <summary>
@@ -99,7 +111,7 @@ namespace CoreEx.Http
         /// <summary>
         /// Create the request applying the specified options and args.
         /// </summary>
-        private async Task<HttpRequestMessage> CreateRequestInternalAsync(HttpMethod method, [StringSyntax(StringSyntaxAttribute.Uri)] string requestUri, HttpContent? content, HttpRequestOptions? requestOptions = null, IEnumerable<IHttpArg>? args = null, CancellationToken cancellationToken = default)
+        private async Task<HttpRequestMessage> CreateRequestInternalAsync(HttpMethod method, string requestUri, HttpContent? content, HttpRequestOptions? requestOptions = null, IEnumerable<IHttpArg>? args = null, CancellationToken cancellationToken = default)
         {
             // Replace any format placeholders within request uri.
             requestUri = FormatReplacement(requestUri, args);